create a separate crate for symcc url and commit hash
also contains functions to checkout and build symcc from a build script
