+++ fuzzers/libfuzzer_stb_image_concolic/fuzzer/Cargo.toml
@@ -24,3 +24,4 @@ cc = { version = "1.0", features = ["parallel"] }
 num_cpus = "1.0"
 cmake = "0.1"
 which = "4.1"
+symcc_libafl = {path = "../../../libafl_concolic/symcc_libafl"}
