+++ libafl_concolic/symcc_libafl/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "symcc_libafl"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[features]
+default = ["build", "clone"]
+build = ["which", "cmake"] # common functionality to build symcc
+clone = ["which"] # common functionality to check out the symcc repo using git
+
+[dependencies]
+which = { version = "4.1", optional = true }
+cmake = { version = "0.1", optional = true }
